# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Blood Bank API",
    "description": "API for the Blood Bank project",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/bloodbank/create": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Bloodbank",
        "description": "Creates a new bloodbank with the logged in admin user as an administrator",
        "operationId": "create_bloodbank",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBloodBankRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/bloodbank/{bloodbank_id}/admin/add": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Add Admin To Bloodbank",
        "description": "Add another admin to a blood bank the logged in user is administrator for",
        "operationId": "add_admin_to_bloodbank",
        "parameters": [
          {
            "name": "bloodbank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bloodbank Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutateBloodbankAdminsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/bloodbank/{bloodbank_id}/admin/remove": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Remove Admin From Bloodbank",
        "description": "Remove an admin from a blood bank the logged in user is administrator for\n\nThe user can remove themselves.\nThe user can't remove the last admin of a blood bank.",
        "operationId": "remove_admin_from_bloodbank",
        "parameters": [
          {
            "name": "bloodbank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bloodbank Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MutateBloodbankAdminsRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/bloodbank/{bloodbank_id}/appointment": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Bloodbank Appointments",
        "description": "Show appointments for blood bank.\n\nBy default only the ones scheduled after 00:00 in Oslo on the same day",
        "operationId": "bloodbank_appointments",
        "parameters": [
          {
            "name": "bloodbank_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bloodbank Id"
            }
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "After"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          },
          {
            "name": "show_cancelled",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Show Cancelled"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingSlotType"
                  },
                  "title": "Response Bloodbank Appointments Admin Bloodbank  Bloodbank Id  Appointment Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/appointment/{appointment_id}/note": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Appointment Add Note",
        "operationId": "appointment_add_note",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/appointment/{appointment_id}/donation": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Appointment Register Donation",
        "operationId": "appointment_register_donation",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDonationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/donor/{donor_id}/form/interview": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Register Interview",
        "operationId": "register_interview",
        "parameters": [
          {
            "name": "donor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Donor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInterviewRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/appointment/{appointment_id}/result": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get Testresult For Appointment Admin",
        "operationId": "get_testresult_for_appointment_admin",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Donation"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/donor/{donor_id}/form/donation-test": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Register Donation Test",
        "operationId": "register_donation_test",
        "parameters": [
          {
            "name": "donor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Donor Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDonationTestRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/appointment/{appointment_id}/note": {
      "post": {
        "summary": "Add Note",
        "operationId": "add_note",
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddNoteRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AddNoteRequest"
      },
...
      "AppointmentTypeAdmin": {
        "properties": {
          "appointment_id": {
            "type": "integer",
            "title": "Appointment Id"
          },
          "appointment_cancelled": {
            "type": "boolean",
            "title": "Appointment Cancelled"
          },
          "donor_id": {
            "type": "integer",
            "title": "Donor Id"
          },
          "donor_blood_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BloodType"
              },
              {
                "type": "null"
              }
            ]
          },
          "donor_name": {
            "type": "string",
            "title": "Donor Name"
          },
          "donor_email": {
            "type": "string",
            "title": "Donor Email"
          },
          "donor_phone": {
            "type": "string",
            "title": "Donor Phone"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteType"
            },
            "type": "array",
            "title": "Notes"
          },
          "donations": {
            "items": {
              "$ref": "#/components/schemas/DonationType"
            },
            "type": "array",
            "title": "Donations"
          }
        },
        "type": "object",
        "required": [
          "appointment_id",
          "appointment_cancelled",
          "donor_id",
          "donor_blood_type",
          "donor_name",
          "donor_email",
          "donor_phone",
          "notes",
          "donations"
        ],
        "title": "AppointmentTypeAdmin"
      },
...
      "BloodType": {
        "type": "string",
        "enum": [
          "O+",
          "O-",
          "A+",
          "A-",
          "B+",
          "B-",
          "AB+",
          "AB-"
        ],
        "title": "BloodType"
      },
...
      "BookingSlotType": {
        "properties": {
          "bookingslot_id": {
            "type": "integer",
            "title": "Bookingslot Id"
          },
          "bookingslot_time": {
            "type": "string",
            "format": "date-time",
            "title": "Bookingslot Time"
          },
          "bookingslot_duration": {
            "type": "string",
            "format": "duration",
            "title": "Bookingslot Duration"
          },
          "bookingslot_remaining_capacity": {
            "type": "integer",
            "title": "Bookingslot Remaining Capacity"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentTypeAdmin"
            },
            "type": "array",
            "title": "Appointments"
          }
        },
        "type": "object",
        "required": [
          "bookingslot_id",
          "bookingslot_time",
          "bookingslot_duration",
          "bookingslot_remaining_capacity",
          "appointments"
        ],
        "title": "BookingSlotType"
      },
      "CreateBloodBankRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "street_name": {
            "type": "string",
            "title": "Street Name"
          },
          "street_number": {
            "type": "string",
            "title": "Street Number"
          },
          "postal_code": {
            "type": "string",
            "title": "Postal Code"
          },
          "city": {
            "type": "string",
            "title": "City"
          },
          "country": {
            "type": "string",
            "title": "Country",
            "default": "NO"
          },
          "loc_lat": {
            "type": "number",
            "title": "Loc Lat"
          },
          "loc_lon": {
            "type": "number",
            "title": "Loc Lon"
          }
        },
        "type": "object",
        "required": [
          "name",
          "street_name",
          "street_number",
          "postal_code",
          "city",
          "loc_lat",
          "loc_lon"
        ],
        "title": "CreateBloodBankRequest"
      },
      "Donation": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "appointment_id": {
            "type": "integer",
            "title": "Appointment Id"
          },
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma"
          }
        },
        "type": "object",
        "required": [
          "id",
          "appointment_id",
          "amount_ml",
          "is_blood_not_plasma"
        ],
        "title": "Donation"
      },
...
      "DonationType": {
        "properties": {
          "donation_id": {
            "type": "integer",
            "title": "Donation Id"
          },
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma"
          }
        },
        "type": "object",
        "required": [
          "donation_id",
          "amount_ml",
          "is_blood_not_plasma"
        ],
        "title": "DonationType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MutateBloodbankAdminsRequest": {
        "properties": {
          "admin_id": {
            "type": "integer",
            "title": "Admin Id"
          }
        },
        "type": "object",
        "required": [
          "admin_id"
        ],
        "title": "MutateBloodbankAdminsRequest"
      },
      "NoteType": {
        "properties": {
          "author_user_id": {
            "type": "integer",
            "title": "Author User Id"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "time": {
            "type": "string",
            "format": "date-time",
            "title": "Time"
          }
        },
        "type": "object",
        "required": [
          "author_user_id",
          "author_name",
          "message",
          "time"
        ],
        "title": "NoteType"
      },
      "RegisterDonationRequest": {
        "properties": {
          "amount_ml": {
            "type": "number",
            "title": "Amount Ml"
          },
          "is_blood_not_plasma": {
            "type": "boolean",
            "title": "Is Blood Not Plasma",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "amount_ml"
        ],
        "title": "RegisterDonationRequest"
      },
      "RegisterDonationTestRequest": {
        "properties": {
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ok_to_donate": {
            "type": "boolean",
            "title": "Ok To Donate"
          },
          "validity_duration": {
            "type": "string",
            "format": "duration",
            "title": "Validity Duration"
          },
          "donation_id": {
            "type": "integer",
            "title": "Donation Id"
          }
        },
        "type": "object",
        "required": [
          "ok_to_donate",
          "validity_duration",
          "donation_id"
        ],
        "title": "RegisterDonationTestRequest"
      },
      "RegisterInterviewRequest": {
        "properties": {
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ok_to_donate": {
            "type": "boolean",
            "title": "Ok To Donate"
          },
          "validity_duration": {
            "type": "string",
            "format": "duration",
            "title": "Validity Duration"
          }
        },
        "type": "object",
        "required": [
          "ok_to_donate",
          "validity_duration"
        ],
        "title": "RegisterInterviewRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}